Fixed attachment delete and file parameters during the draft attachment download

+++ modules/smtp/modules.php
@@ -58,14 +58,14 @@ class Hm_Handler_load_smtp_is_imap_draft extends Hm_Handler_Module {
                 
                 $attached_files = [];
                 $this->session->set('uploaded_files', array());
-                if (array_key_exists('subs', $msg_struct[0])) {
+                if (array_key_exists(0, $msg_struct) && array_key_exists('subs', $msg_struct[0])) {
                     foreach ($msg_struct[0]['subs'] as $ind => $sub) {
                         if ($ind != '0.1') {
-                            $new_attachment['basename'] = $sub['attributes']['name'];
-                            $new_attachment['name'] = $sub['attributes']['name'];
+                            $new_attachment['basename'] = $sub['description'];
+                            $new_attachment['name'] = $sub['description'];
                             $new_attachment['size'] = $sub['size'];
                             $new_attachment['type'] = $sub['type'];
-                            $file_path = $this->config->get('attachment_dir').DIRECTORY_SEPARATOR.$sub['attributes']['name'];
+                            $file_path = $this->config->get('attachment_dir').DIRECTORY_SEPARATOR.$new_attachment['name'];
                             $content = Hm_Crypt::ciphertext($imap->get_message_content($this->request->get['uid'], $ind), Hm_Request_Key::generate());
                             file_put_contents($file_path, $content);
                             $new_attachment['tmp_name'] = $file_path;
@@ -881,6 +881,7 @@ class Hm_Output_compose_form_content extends Hm_Output_Module {
                 '<script type="text/javascript">var editor = new Editor(); editor.render();</script>';
         }
         $res .= '<table class="uploaded_files">';
+        
         foreach ($files as $file) {
             $res .= format_attachment_row($file, $this);
         }
